Fixes for the following issues:
One run of  python file_organizer.py -R --scan-once filled up my Google Drive. It put thousands of files in Trash (bin) in the drive, and lots of small files in ~/Google\ Drive/My\ Drive/MyFiles, which were in various subfolders of ~/dev/ on my main drive. I don't know how either of these things happened, but more important is running out of disk space. It would be a great improvement if every time a folder is being copied to a drive, it is tested for being less than 90% full, if that's possible.

The next issue is the soft link folders being created in ~/organized. It doesn't create three-letter ones like 'eae' any more, but it creates lots of four-letter ones with equally meaningless names. Again, the linked to by soft links in these soft link folders, do not contain those words in their names. For example, ~/organized/meee contains vidi-washing-machine.jpg - and so do eight other meaningless four-letter folders. Despite the profligacy of links to this file, there is no folder ~/organized/vidi.

Neither is there a folder ~/organized/vienna, despite there being numerous links in numerous soft link folders to files with 'vienna' in their names. E.g. 31-05-18-vienna-shoppe-kirche.jpg.

The only references to these meaningless four-letter words in the log are, e.g.
2025-10-10 11:01:24,808 - file_organizer - INFO - Discovered category 'meee' with 115 files (keyword frequency: 115)

In the log, it doesn't say it's finding these four-letter not-words in images, so I don't know where these words are coming from. OCR finds some odd things in the images, but not 'meee', 'saat', etc.

2025-10-10 10:52:39,499 - file_organizer - INFO - EasyOCR found: 7 UTHEP 11 LACE...
2025-10-10 10:52:45,814 - file_organizer - INFO - EasyOCR found: Rather than flood out of the ground, the home fans...
2025-10-10 10:52:51,898 - file_organizer - INFO - EasyOCR found: 170n304917575 3 KM: 4 ACCOMMODATION SERVICE 3 KM C...
2025-10-10 10:53:10,548 - file_organizer - INFO - EasyOCR found: PHRA THAT doi SUTHEP 13 BHUBING PALACE...
2025-10-10 10:53:19,306 - file_organizer - INFO - EasyOCR found: 1 8 1 2 0 2 8 2 8 8 7 ( 1 8 2 X 2 7 2 U 2 8 9 8 1 ...
2025-10-10 10:53:29,617 - file_organizer - INFO - EasyOCR found: 8 6 8 4 3 5 5 1 3 8 J 1 1 9 ;...
2025-10-10 10:53:47,312 - file_organizer - INFO - EasyOCR found: 9 1exieieuea 1 eqJonIn ZVOZ31 alivir...

I don't want it to consult a dictionary, but I definitely want it to isolate words separated by dashes, underscores, or dots, or at the beginning or end of filenames.

For a human, it's easy to see that eqJonIn is not a word, but alivir might be. And ZVOZ31 is probably random noise. PHRA and SUTHEP are Thai. I would suggest that subfolder names based on words found in images over four letters long would be given subfolders, lower-case.

The partial sentence "Rather than flood out of the ground, the home fans..." found in an image tells me immediately that a soft link to this file should go in a soft folder 'football', but I can't expect the program to know that. If it put it in folders 'rather' and 'flood' it would be doing OK. So... only bother with words above four letters long found in image files, and only if they are all characters, no numbers.

From the above example OCR output, I would expect the following subfolders
eqjonin
alivir
rather
flood
suthep

One more thing. The file 31-05-18-vienna-shoppe-kirche.jpg isn't YYYYMMDD format.
It's also not American date format. If possible, I'd have it interpret
01-05-18 as 2018-05-01
05-31-18 as 2018-05-31
in other words "don't go American unless you have to".